requirements.txt
hydra-core
rich
pydantic
scikit-learn
pandas
numpy
datasets
langchain
python-dotenv
langchain-openai
accelerate
sentence-transformers
torch
matplotlib
